requirements.txt
appdirs==1.4.4
black==21.7b0
cffi==1.14.6
click==8.0.1
configobj==5.0.6
cryptography==3.4.7
future==0.18.2
install==1.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyOpenSSL==20.0.1
python-gnupg==0.4.7
python-ldap==3.3.1
regex==2021.7.6
secrets==1.0.2
six==1.16.0
systematic==4.8.7
tomli==1.2.0
versioneer==0.20
